+++ include/yams/compression/compressor_interface.h
@@ -40,7 +40,9 @@ struct CompressionResult {
      * @return Ratio of original to compressed size
      */
     [[nodiscard]] double ratio() const noexcept {
-        return compressedSize > 0 ? static_cast<double>(originalSize) / compressedSize : 0.0;
+        return compressedSize > 0
+                   ? static_cast<double>(originalSize) / static_cast<double>(compressedSize)
+                   : 0.0;
     }
 
     /**
@@ -48,8 +50,10 @@ struct CompressionResult {
      * @return Percentage of space saved
      */
     [[nodiscard]] double spaceSaved() const noexcept {
-        return originalSize > 0 ? (1.0 - static_cast<double>(compressedSize) / originalSize) * 100.0
-                                : 0.0;
+        return originalSize > 0
+                   ? (1.0 - static_cast<double>(compressedSize) / static_cast<double>(originalSize)) *
+                         100.0
+                   : 0.0;
     }
 };
